Remove redundant code

+++ WinQuake/pr_edict.c
@@ -743,7 +743,7 @@ qboolean	ED_ParseEpair (void *base, ddef_t *key, char *s)
 		break;
 		
 	case ev_float:
-		*(float *)d = atof (s);
+		*(float *)d = (float)atof (s);
 		break;
 		
 	case ev_vector:
@@ -755,7 +755,7 @@ qboolean	ED_ParseEpair (void *base, ddef_t *key, char *s)
 			while (*v && *v != ' ')
 				v++;
 			*v = 0;
-			((float *)d)[i] = atof (w);
+			((float *)d)[i] = (float)atof (w);
 			w = v = v+1;
 		}
 		break;
